[FileUploader] UI/UX improvements

+++ wegas-app/src/main/webapp/wegas-app/js/widget/wegas-file-uploader.js
@@ -5,21 +5,23 @@
  * Copyright (c) 2013-2020 School of Business and Engineering Vaud, Comem, MEI
  * Licensed under the MIT License
  */
+/* global I18n */
+
 YUI.add('wegas-file-uploader', function(Y) {
     'use strict';
-    var CONTENTBOX = 'contentBox';
+
     var FileUploader = Y.Base.create('wegas-file-uploader', Y.Widget,
         [Y.WidgetChild, Y.Wegas.Widget, Y.Wegas.Editable], {
         CONTENT_TEMPLATE: ""
             + "<div>"
-            + "   <div class='wegas-file-uploader--uploader'>"
-            + "    <label>"
-            + "      <span class='wegas-file-uploader--input-label'></span>"
-            + "      <input class='wegas-file-uploader--input-input' style='display: none' type='file' />"
-            + "    </label>"
+            + "  <div class='wegas-file-uploader--dropzone-container'>"
+            + "    <div class='wegas-file-uploader--dropzone'></div>"
             + "  </div>"
-            + "  <div class='wegas-file-uploader--filename'>"
-            + "  </div>"
+            + "  <div class='wegas-file-uploader--filename'></div>"
+            + "  <label class='wegas-file-uploader--input'>"
+            + "    <div class='wegas-file-uploader--input-label'></div>"
+            + "    <input class='wegas-file-uploader--input-input' style='display: none' type='file' />"
+            + "  </label>"
             + "</div>"
         ,
         initializer: function() {
@@ -33,6 +35,7 @@ YUI.add('wegas-file-uploader', function(Y) {
             }
         },
         renderUI: function() {
+            // no op
         },
         bindUI: function() {
             var instance = this.get("shortcutVar.evaluated").getInstance();
@@ -40,45 +43,83 @@ YUI.add('wegas-file-uploader', function(Y) {
                 this.handlers.onUpdate = Y.Wegas.Facade.Instance.after(
                     instance.get("id") + ':updatedInstance', this.syncUI, this);
             }
+            var BB = this.get("boundingBox");
             var CB = this.get("contentBox");
-            this.handlers.onRootDragOver = this.get("root").on("dragover", function(e) {
-            }, this);
-            this.handlers.onRootDrop = this.get("root").get("boundingBox").on("drop", function(e) {
+            //this.handlers.onRootDragOver = this.get("root").on("dragover", function(e) {
+            //}, this);
+
+            this.get("root").get("boundingBox").on('drop', function(e) {
+                e.halt(true);
+            });
+            //Prevent drop to avoid application exit
+            this.get("root").get("boundingBox").on('dragover', function(e) {
+//                e._event.dataTransfer.dropEffect = 'none';
+                e.halt(true);
+            });
+
+            // on drop -> upload
+            BB.delegate("drop", function(e) {
                 e.preventDefault();
                 e.halt(true);
-                // prevent drop to avoid application exit
-            }, this);
-            CB.delegate("drop", function(e) {
-                e.preventDefault();
+                this.get("boundingBox").removeClass("highlight");
+                if (e._event.dataTransfer.files.length) {
+                    this.upload(e._event.dataTransfer.files[0]);
+                }
+                // todo check accept
+            }, ".wegas-file-uploader--dropzone", this);
+
+            BB.delegate("dragover", function(e) {
                 e.halt(true);
-            }, ".wegas-file-uploader--uploader", this);
-            CB.delegate("dragenter", function(e) {
-                e.target.addClass("highlight");
-            }, ".wegas-file-uploader--uploader", this);
-            CB.delegate("dragleave", function(e) {
-                e.target.removeClass("highlight");
-            }, ".wegas-file-uploader--uploader", this);
+            }, ".wegas-file-uploader--dropzone", this);
+
+
+            // highlight on hover
+            BB.delegate("dragenter", function(e) {
+                var bb = this.get("boundingBox");
+                bb.addClass("highlight");
+
+                // dataTransfer is not available on dragenter
+//                if (e._event.dataTransfer.files.length && !this.isExtensionValid(e._event.dataTransfer.files[0])) {
+//                    bb.addClass("invalid-filetype");
+//                }
+            }, ".wegas-file-uploader--dropzone", this);
+
+            // un highlight
+            BB.delegate("dragleave", function(e) {
+                var bb = this.get("boundingBox");
+                bb.removeClass("highlight");
+//                bb.removeClass("invalid-filetype");
+            }, ".wegas-file-uploader--dropzone", this);
+
             CB.delegate("change", function(e) {
                 this.upload(e.target.getDOMNode().files[0]);
             }, ".wegas-file-uploader--input-input", this);
+
             //this.handlers.onSelectFile = CB.delegate("cl");
         },
-        getUploadLabel: function() {
-            return this.get("uploadLabel");
+        isExtensionValid: function(file) {
+            var filename = file.name;
+            return !!this.get("accept").split(",").find(function(ext) {
+                return filename.endsWith(ext.trim());
+            });
         },
-        getUpdateLabel: function() {
-            return this.get("updateLabel");
+        getUploadLabel: function() {
+            var labelVar = this.get("uploadLabelVar.evaluated");
+            if (labelVar) {
+                return I18n.tVar(labelVar);
+            } else {
+                return "upload";
+            }
         },
         syncUI: function() {
             var data = this.readLink();
             var theLabel = this.get("contentBox").one(".wegas-file-uploader--input-label");
             var theInput = this.get("contentBox").one(".wegas-file-uploader--input-input");
+
             theInput.getDOMNode().accept = this.get("accept");
-            if (data && data.filename) {
-                theLabel.getDOMNode().innerText = this.getUpdateLabel();
-            } else {
-                theLabel.getDOMNode().innerText = this.getUploadLabel();
-            }
+
+            theLabel.getDOMNode().innerHTML = this.getUploadLabel();
+
             this._updateLink(data.link);
         },
         _updateLink: function(link) {
@@ -88,8 +129,8 @@ YUI.add('wegas-file-uploader', function(Y) {
             //theLink.getDOMNode().innerText = data.displayName;
         },
         _addSlashes: function(oPath) {
-            var path = oPath;
-            if (path.charAt(0) !== '/') {
+            var path = oPath || "";
+            if (path.length === 0 || path.charAt(0) !== '/') {
                 path = "/" + path;
             }
             if (!path.endsWith("/")) {
@@ -160,11 +201,11 @@ YUI.add('wegas-file-uploader', function(Y) {
                         var varName = this.get("shortcutVar.evaluated").get('name');
 
                         if (response.ok) {
-                            var link = "<a target='_blank' data-file='"
+                            var link = '<a target="_blank" data-file=\"'
                                 + path + file.name
-                                + "'>"
+                                + '">'
                                 + file.name// display name
-                                + "</a>";
+                                + '</a>';
                             Y.Wegas.Facade.Variable.script.remoteEval(
                                 'Variable.find(gameModel, "' + varName + '")'
                                 + '.setValue(self, ' + JSON.stringify(link) + ');');
@@ -172,7 +213,7 @@ YUI.add('wegas-file-uploader', function(Y) {
                         } else {
                             response.json().then(function(error) {
                                 var content = "Fails to uplaod file";
-                                if (error && error.message){
+                                if (error && error.message) {
                                     content += ": " + error.message
                                 }
                                 Y.Wegas.Facade.Variable.script.remoteEval(
@@ -189,28 +230,32 @@ YUI.add('wegas-file-uploader', function(Y) {
         upload: function(file) {
 
             if (file) {
-                var currentFile = this.readLink();
-                if (currentFile && currentFile.filename) {
-                    // Delete existing file
-                    var gmId = Y.Wegas.Facade.GameModel.cache.getCurrentGameModel().get("id");
-                    var url = '/' + gmId + '/File/delete' + currentFile.filename;
-                    Y.Wegas.Facade.GameModel.sendRequest({
-                        request: url,
-                        cfg: {
-                            method: 'DELETE'
-                        },
-                        on: {
-                            success: Y.bind(function(_e) {
-                                this._processUpload(file);
-                            }, this),
-                            failure: Y.bind(function(_e) {
-                                // Error
-                                this._processUpload(file);
-                            }, this)
-                        }
-                    });
+                if (this.isExtensionValid(file)) {
+                    var currentFile = this.readLink();
+                    if (currentFile && currentFile.filename) {
+                        // Delete existing file
+                        var gmId = Y.Wegas.Facade.GameModel.cache.getCurrentGameModel().get("id");
+                        var url = '/' + gmId + '/File/delete' + currentFile.filename;
+                        Y.Wegas.Facade.GameModel.sendRequest({
+                            request: url,
+                            cfg: {
+                                method: 'DELETE'
+                            },
+                            on: {
+                                success: Y.bind(function(_e) {
+                                    this._processUpload(file);
+                                }, this),
+                                failure: Y.bind(function(_e) {
+                                    // Error
+                                    this._processUpload(file);
+                                }, this)
+                            }
+                        });
+                    } else {
+                        this._processUpload(file);
+                    }
                 } else {
-                    this._processUpload(file);
+                    Y.Wegas.Alerts.showMessage('error', "Invalid file type");
                 }
             } else {
                 // no-op
@@ -221,7 +266,15 @@ YUI.add('wegas-file-uploader', function(Y) {
          * @returns {}
          */
         readLink: function() {
-            var theVar = this.get("shortcutVar.evaluated");
+            return FileUploader.readLink(this.get("shortcutVar.evaluated"));
+        }
+    }, {
+        /**
+         * read HTML encoded link (<a>...</a>) from a String/Text Descriptor
+         *
+         * EG: <p><a data-file="/logos/MB_200.png">Pres.pptx</a></p>
+         */
+        readLink: function(theVar) {
             var ret = {
                 variable: theVar,
                 displayName: '',
@@ -245,8 +298,7 @@ YUI.add('wegas-file-uploader', function(Y) {
                 }
             }
             return ret;
-        }
-    }, {
+        },
         EDITORNAME: 'Uploader',
         ATTRS: {
             /**
@@ -264,23 +316,23 @@ YUI.add('wegas-file-uploader', function(Y) {
             },
             accept: {
                 type: "string",
+                value: "",
                 view: {
                     "label": "Accept",
                     "description": "File extension to accept, eg \".jpg, .png\""
                 }
             },
-            uploadLabel: {
-                type: "string",
-                value: "Upload",
+            uploadLabelVar: {
+                type: "object",
+                getter: Y.Wegas.Widget.VARIABLEDESCRIPTORGETTER,
                 view: {
-                    "label": "Empty label",
-                }
-            },
-            updateLabel: {
-                type: "string",
-                value: "Update",
-                view: {
-                    "label": "Update label",
+                    type: 'variableselect',
+                    label: 'Upload Button Label',
+                    classFilter: [
+                        'TextDescriptor',
+                        'StaticTextDescriptor',
+                        'StringDescriptor'
+                    ]
                 }
             },
             directory: {
